+++ crates/holochain/src/conductor/state.rs
@@ -266,15 +266,25 @@ pub struct AppInterfaceConfig {
     /// The signal subscription settings for each App
     pub signal_subscriptions: HashMap<InstalledAppId, SignalSubscription>,
 
+    /// The application that this interface is for. If `Some`, then this interface will only allow
+    /// connections which use a token that has been issued for the same app id. Otherwise, any app
+    /// is allowed to connect.
+    pub installed_app_id: Option<InstalledAppId>,
+
     /// The driver for the interface, e.g. Websocket
     pub driver: InterfaceDriver,
 }
 
 impl AppInterfaceConfig {
     /// Create config for a websocket interface
-    pub fn websocket(port: u16, allowed_origins: AllowedOrigins) -> Self {
+    pub fn websocket(
+        port: u16,
+        allowed_origins: AllowedOrigins,
+        installed_app_id: Option<InstalledAppId>,
+    ) -> Self {
         Self {
             signal_subscriptions: HashMap::new(),
+            installed_app_id,
             driver: InterfaceDriver::Websocket {
                 port,
                 allowed_origins,