+++ crates/holochain/tests/websocket/mod.rs
@@ -3,9 +3,9 @@ use anyhow::Result;
 use futures::future;
 use hdk::prelude::RemoteSignal;
 use holochain::conductor::interface::websocket::MAX_CONNECTIONS;
-use holochain::sweettest::SweetConductor;
 use holochain::sweettest::SweetConductorBatch;
 use holochain::sweettest::SweetDnaFile;
+use holochain::sweettest::{authenticate_app_ws_client, SweetConductor, WsPollRecv};
 use holochain::sweettest::{SweetAgents, SweetConductorConfig};
 use holochain::{
     conductor::{
@@ -15,9 +15,14 @@ use holochain::{
     },
     fixt::*,
 };
-use std::net::ToSocketAddrs;
+use std::net::{Ipv4Addr, Ipv6Addr, ToSocketAddrs};
 
-use holochain_conductor_api::{AdminInterfaceConfig, AppRequest, InterfaceDriver};
+use either::Either;
+use holochain_conductor_api::{
+    AdminInterfaceConfig, AppAuthenticationRequest, AppAuthenticationToken, AppRequest,
+    InterfaceDriver, IssueAppAuthenticationTokenPayload,
+};
+use holochain_types::websocket::AllowedOrigins;
 use holochain_types::{
     prelude::*,
     test_utils::{fake_dna_zomes, write_fake_dna_file},
@@ -32,26 +37,6 @@ use tempfile::TempDir;
 use tracing::*;
 
 use crate::test_utils::*;
-
-struct PollRecv(tokio::task::JoinHandle<()>);
-
-impl Drop for PollRecv {
-    fn drop(&mut self) {
-        self.0.abort();
-    }
-}
-
-impl PollRecv {
-    pub fn new<D>(mut rx: WebsocketReceiver) -> Self
-    where
-        D: std::fmt::Debug,
-        SerializedBytes: TryInto<D, Error = SerializedBytesError>,
-    {
-        Self(tokio::task::spawn(async move {
-            while rx.recv::<D>().await.is_ok() {}
-        }))
-    }
-}
 
 #[tokio::test(flavor = "multi_thread")]
 #[cfg(feature = "slow_tests")]
@@ -78,7 +63,7 @@ async fn call_admin() {
     let port = port.await.unwrap();
 
     let (mut client, rx) = websocket_client_by_port(port).await.unwrap();
-    let _rx = PollRecv::new::<AdminResponse>(rx);
+    let _rx = WsPollRecv::new::<AdminResponse>(rx);
 
     // Make properties
     let properties = holochain_zome_types::properties::YamlProperties::new(
@@ -136,7 +121,7 @@ async fn call_zome() {
     let admin_port = admin_port.await.unwrap();
 
     let (mut admin_tx, admin_rx) = websocket_client_by_port(admin_port).await.unwrap();
-    let _admin_rx = PollRecv::new::<AdminResponse>(admin_rx);
+    let _admin_rx = WsPollRecv::new::<AdminResponse>(admin_rx);
     let (_, mut receiver2) = websocket_client_by_port(admin_port).await.unwrap();
 
     let uuid = uuid::Uuid::new_v4();
@@ -154,7 +139,7 @@ async fn call_zome() {
     // List Dnas
     let request = AdminRequest::ListDnas;
     let response = admin_tx.request(request);
-    let response = check_timeout(response, 3000).await;
+    let response = check_timeout(response, 15000).await;
 
     assert_matches!(response, AdminResponse::DnasListed(a) if a.contains(&installed_dna_hash));
 
@@ -187,7 +172,8 @@ async fn call_zome() {
     let app_port = attach_app_interface(&mut admin_tx, None).await;
 
     let (mut app_tx, app_rx) = websocket_client_by_port(app_port).await.unwrap();
-    let _app_rx = PollRecv::new::<AppResponse>(app_rx);
+    let _app_rx = WsPollRecv::new::<AppResponse>(app_rx);
+    authenticate_app_ws_client(app_tx.clone(), admin_port, "test".to_string()).await;
 
     // Call Zome
     tracing::info!("Calling zome");
@@ -223,7 +209,7 @@ async fn call_zome() {
     let admin_port = admin_port.await.unwrap();
 
     let (admin_tx, admin_rx) = websocket_client_by_port(admin_port).await.unwrap();
-    let _admin_rx = PollRecv::new::<AdminResponse>(admin_rx);
+    let _admin_rx = WsPollRecv::new::<AdminResponse>(admin_rx);
 
     tokio::time::sleep(std::time::Duration::from_millis(1000)).await;
 
@@ -231,12 +217,13 @@ async fn call_zome() {
     let response = admin_tx.request(request);
     let response = check_timeout(response, 3000).await;
     let app_port = match response {
-        AdminResponse::AppInterfacesListed(ports) => *ports.first().unwrap(),
+        AdminResponse::AppInterfacesListed(ports) => ports.first().map(|i| i.port).unwrap(),
         _ => panic!("Unexpected response"),
     };
 
     let (app_tx, app_rx) = websocket_client_by_port(app_port).await.unwrap();
-    let _app_rx = PollRecv::new::<AppResponse>(app_rx);
+    let _app_rx = WsPollRecv::new::<AppResponse>(app_rx);
+    authenticate_app_ws_client(app_tx.clone(), admin_port, "test".to_string()).await;
 
     // Call Zome again on the existing app interface port
     tracing::info!("Calling zome again");
@@ -344,7 +331,7 @@ async fn emit_signals() {
     let admin_port = admin_port.await.unwrap();
 
     let (mut admin_tx, admin_rx) = websocket_client_by_port(admin_port).await.unwrap();
-    let _admin_rx = PollRecv::new::<AdminResponse>(admin_rx);
+    let _admin_rx = WsPollRecv::new::<AdminResponse>(admin_rx);
 
     let uuid = uuid::Uuid::new_v4();
     let dna = fake_dna_zomes(
@@ -403,8 +390,9 @@ async fn emit_signals() {
             }
         }
     });
+    authenticate_app_ws_client(app_tx_1.clone(), admin_port, "test".to_string()).await;
 
-    let (_, mut app_rx_2) = websocket_client_by_port(app_port).await.unwrap();
+    let (app_tx_2, mut app_rx_2) = websocket_client_by_port(app_port).await.unwrap();
     let (sig2_send, sig2_recv) = tokio::sync::oneshot::channel();
     let mut sig2_send = Some(sig2_send);
     let sig2_task = tokio::task::spawn(async move {
@@ -419,6 +407,7 @@ async fn emit_signals() {
             }
         }
     });
+    authenticate_app_ws_client(app_tx_2.clone(), admin_port, "test".to_string()).await;
 
     call_zome_fn(
         &app_tx_1,
@@ -457,7 +446,7 @@ async fn conductor_admin_interface_runs_from_config() -> Result<()> {
     let config = create_config(0, environment_path.into());
     let conductor_handle = Conductor::builder().config(config).build().await?;
     let (client, rx) = websocket_client(&conductor_handle).await?;
-    let _rx = PollRecv::new::<AdminResponse>(rx);
+    let _rx = WsPollRecv::new::<AdminResponse>(rx);
 
     let dna = fake_dna_zomes("", vec![(TestWasm::Foo.into(), TestWasm::Foo.into())]);
     let (fake_dna_path, _tmpdir) = write_fake_dna_file(dna).await.unwrap();
@@ -498,7 +487,7 @@ async fn list_app_interfaces_succeeds() -> Result<()> {
         ),
     )
     .await?;
-    let _rx = PollRecv::new::<AdminResponse>(rx);
+    let _rx = WsPollRecv::new::<AdminResponse>(rx);
 
     let request = AdminRequest::ListAppInterfaces;
 
@@ -611,7 +600,7 @@ async fn connection_limit_is_respected() {
     // The first `MAX_CONNECTIONS` connections should succeed
     for count in 0..MAX_CONNECTIONS {
         let (sender, rx) = connect(cfg.clone(), addr).await.unwrap();
-        let rx = PollRecv::new::<AdminResponse>(rx);
+        let rx = WsPollRecv::new::<AdminResponse>(rx);
         let _: AdminResponse = sender
             .request(AdminRequest::ListDnas)
             .await
@@ -623,7 +612,7 @@ async fn connection_limit_is_respected() {
     // Try lots of failed connections to make sure the limit is respected
     for _ in 0..2 * MAX_CONNECTIONS {
         let (sender, rx) = connect(cfg.clone(), addr).await.unwrap();
-        let _rx = PollRecv::new::<AdminResponse>(rx);
+        let _rx = WsPollRecv::new::<AdminResponse>(rx);
 
         // Getting a sender back isn't enough to know that the connection succeeded because the other side takes a moment to shutdown, try sending to be sure
         sender
@@ -638,7 +627,7 @@ async fn connection_limit_is_respected() {
     // Should now be possible to connect new clients
     for count in 0..MAX_CONNECTIONS {
         let (sender, rx) = connect(cfg.clone(), addr).await.unwrap();
-        let rx = PollRecv::new::<AdminResponse>(rx);
+        let rx = WsPollRecv::new::<AdminResponse>(rx);
         let _: AdminResponse = sender
             .request(AdminRequest::ListDnas)
             .await
@@ -676,7 +665,7 @@ async fn concurrent_install_dna() {
     let admin_port = admin_port.await.unwrap();
 
     let (client, rx) = websocket_client_by_port(admin_port).await.unwrap();
-    let _rx = PollRecv::new::<AdminResponse>(rx);
+    let _rx = WsPollRecv::new::<AdminResponse>(rx);
 
     // let before = std::time::Instant::now();
 
@@ -739,8 +728,11 @@ async fn network_stats() {
     let _ = batch.setup_app("app", &[dna_file]).await.unwrap();
     batch.exchange_peer_info().await;
 
-    let (client, rx) = batch.get(0).unwrap().admin_ws_client().await;
-    let _rx = PollRecv::new::<AdminResponse>(rx);
+    let (client, _rx) = batch
+        .get(0)
+        .unwrap()
+        .admin_ws_client::<AdminResponse>()
+        .await;
 
     #[cfg(feature = "tx5")]
     const EXPECT: &str = "go-pion";
@@ -773,8 +765,7 @@ async fn full_state_dump_cursor_works() {
 
     let cell_id = app.into_cells()[0].cell_id().clone();
 
-    let (mut client, rx) = conductor.admin_ws_client().await;
-    let _rx = PollRecv::new::<AdminResponse>(rx);
+    let (mut client, _rx) = conductor.admin_ws_client::<AppResponse>().await;
 
     let full_state = dump_full_state(&mut client, cell_id.clone(), None).await;
 
@@ -851,7 +842,7 @@ async fn admin_allowed_origins() {
     .await
     .unwrap();
 
-    let _rx = PollRecv::new::<AdminResponse>(rx);
+    let _rx = WsPollRecv::new::<AdminResponse>(rx);
 
     let request = AdminRequest::ListAppInterfaces;
     let _: AdminResponse = client.request(request).await.unwrap();
@@ -868,6 +859,7 @@ async fn app_allowed_origins() {
         .add_app_interface(
             either::Either::Left(0),
             "http://localhost:3000".to_string().into(),
+            None,
         )
         .await
         .unwrap();
@@ -885,7 +877,9 @@ async fn app_allowed_origins() {
     .await
     .is_err());
 
-    check_app_port(port, "http://localhost:3000").await;
+    let token = create_multi_use_token(&conductor).await;
+
+    check_app_port(port, "http://localhost:3000", token).await;
 }
 
 #[tokio::test(flavor = "multi_thread")]
@@ -894,11 +888,14 @@ async fn app_allowed_origins_independence() {
 
     let conductor = SweetConductor::from_standard_config().await;
 
+    let token = create_multi_use_token(&conductor).await;
+
     let port_1 = conductor
         .clone()
         .add_app_interface(
             either::Either::Left(0),
             "http://localhost:3001".to_string().into(),
+            None,
         )
         .await
         .unwrap();
@@ -908,6 +905,7 @@ async fn app_allowed_origins_independence() {
         .add_app_interface(
             either::Either::Left(0),
             "http://localhost:3002".to_string().into(),
+            None,
         )
         .await
         .unwrap();
@@ -946,11 +944,126 @@ async fn app_allowed_origins_independence() {
 
     // Check that correct access is allowed
 
-    check_app_port(port_1, "http://localhost:3001").await;
-    check_app_port(port_2, "http://localhost:3002").await;
+    check_app_port(port_1, "http://localhost:3001", token.clone()).await;
+    check_app_port(port_2, "http://localhost:3002", token).await;
 }
 
-async fn check_app_port(port: u16, origin: &str) {
+async fn create_multi_use_token(conductor: &SweetConductor) -> AppAuthenticationToken {
+    let (admin_sender, _admin_rx) = conductor.admin_ws_client::<AdminResponse>().await;
+
+    let token_response: AdminResponse = admin_sender
+        .request(AdminRequest::IssueAppAuthenticationToken(
+            IssueAppAuthenticationTokenPayload::for_installed_app_id("test_app".into())
+                .single_use(false),
+        ))
+        .await
+        .unwrap();
+    let token = match token_response {
+        AdminResponse::AppAuthenticationTokenIssued(issued) => issued.token,
+        _ => panic!("unexpected response"),
+    };
+
+    token
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn holochain_websockets_listen_on_ipv4_and_ipv6() {
+    holochain_trace::test_run();
+
+    let conductor = SweetConductor::from_standard_config().await;
+
+    let admin_port = conductor.get_arbitrary_admin_websocket_port().unwrap();
+
+    //
+    // Connect to the admin interface on ipv4 and ipv6 localhost
+    //
+
+    let (ipv4_admin_sender, rx) = connect(
+        Arc::new(WebsocketConfig::CLIENT_DEFAULT),
+        ConnectRequest::new((Ipv4Addr::LOCALHOST, admin_port).into()),
+    )
+    .await
+    .unwrap();
+    let _rx4 = WsPollRecv::new::<AdminResponse>(rx);
+
+    let response: AdminResponse = ipv4_admin_sender
+        .request(AdminRequest::ListCellIds)
+        .await
+        .unwrap();
+    match response {
+        AdminResponse::CellIdsListed(_) => (),
+        _ => panic!("unexpected response"),
+    }
+
+    let (ipv6_admin_sender, rx) = connect(
+        Arc::new(WebsocketConfig::CLIENT_DEFAULT),
+        ConnectRequest::new((Ipv6Addr::LOCALHOST, admin_port).into()),
+    )
+    .await
+    .unwrap();
+    let _rx6 = WsPollRecv::new::<AdminResponse>(rx);
+
+    let response: AdminResponse = ipv6_admin_sender
+        .request(AdminRequest::ListCellIds)
+        .await
+        .unwrap();
+    match response {
+        AdminResponse::CellIdsListed(_) => (),
+        _ => panic!("unexpected response"),
+    }
+
+    //
+    // Do the same for an app interface
+    //
+
+    let app_port = conductor
+        .clone()
+        .add_app_interface(Either::Left(0), AllowedOrigins::Any, None)
+        .await
+        .unwrap();
+
+    let (ipv4_app_sender, rx) = connect(
+        Arc::new(WebsocketConfig::CLIENT_DEFAULT),
+        ConnectRequest::new((Ipv4Addr::LOCALHOST, app_port).into()),
+    )
+    .await
+    .unwrap();
+    let _rx4 = WsPollRecv::new::<AppResponse>(rx);
+    authenticate_app_ws_client(ipv4_app_sender.clone(), admin_port, "".to_string()).await;
+
+    let response: AppResponse = ipv4_app_sender
+        .request(AppRequest::AppInfo {
+            installed_app_id: "".to_string(),
+        })
+        .await
+        .unwrap();
+    match response {
+        AppResponse::AppInfo(_) => (),
+        _ => panic!("unexpected response"),
+    }
+
+    let (ipv6_app_sender, rx) = connect(
+        Arc::new(WebsocketConfig::CLIENT_DEFAULT),
+        ConnectRequest::new((Ipv6Addr::LOCALHOST, app_port).into()),
+    )
+    .await
+    .unwrap();
+    let _rx6 = WsPollRecv::new::<AppResponse>(rx);
+    authenticate_app_ws_client(ipv6_app_sender.clone(), admin_port, "".to_string()).await;
+
+    let response: AppResponse = ipv6_app_sender
+        .request(AppRequest::AppInfo {
+            installed_app_id: "".to_string(),
+        })
+        .await
+        .unwrap();
+    match response {
+        AppResponse::AppInfo(_) => (),
+        _ => panic!("unexpected response"),
+    }
+}
+
+async fn check_app_port(port: u16, origin: &str, token: AppAuthenticationToken) {
     let (client, rx) = connect(
         Arc::new(WebsocketConfig::CLIENT_DEFAULT),
         ConnectRequest::new(
@@ -966,7 +1079,12 @@ async fn check_app_port(port: u16, origin: &str) {
     .await
     .unwrap();
 
-    let _rx = PollRecv::new::<AppResponse>(rx);
+    let _rx = WsPollRecv::new::<AppResponse>(rx);
+
+    client
+        .authenticate(AppAuthenticationRequest { token })
+        .await
+        .unwrap();
 
     let request = AppRequest::ListWasmHostFunctions;
     let _: AppResponse = client.request(request).await.unwrap();